# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/restaurants": {
            "get": {
                "description": "Get restaurants",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurants Actions"
                ],
                "summary": "Get restaurants",
                "operationId": "get-restaurants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RestaurantCollectionResponse"
                        }
                    }
                }
            }
        },
        "/restaurants/{id}/menu": {
            "get": {
                "description": "Get menu by restaurant id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant Actions"
                ],
                "summary": "Get restaurant menu position by restaurant id and position id",
                "operationId": "get-restaurant-menu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Restaurant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MenuResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/restaurants/{id}/menu/{position_id}": {
            "get": {
                "description": "Get menu position by restaurant id and position id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant Actions"
                ],
                "summary": "Get restaurant menu by restaurant id and position id",
                "operationId": "get-restaurant-menu-position",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Restaurant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Restaurant menu position ID",
                        "name": "position_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Menu"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Menu": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "response.MenuResponse": {
            "type": "object",
            "properties": {
                "menu": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                }
            }
        },
        "response.RestaurantCollectionResponse": {
            "type": "object",
            "properties": {
                "restaurants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.RestaurantResponse"
                    }
                }
            }
        },
        "response.RestaurantResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}